+++ example/lib/main.dart
@@ -67,8 +67,6 @@ class _AppState extends State<App> {
               hasArrow: true,
               hasLabelBox: true,
               fullscreen: true,
-              manualNextControl: true,
-              closeKey: closeKey,
               stepBuilder: (
                 BuildContext context,
                 OnboardingStepRenderInfo renderInfo,
@@ -106,7 +104,6 @@ class _AppState extends State<App> {
                               child: Text('Next'),
                             ),
                             TextButton(
-                              key: closeKey,
                               onPressed: renderInfo.close,
                               child: Text('close'),
                             ),
@@ -146,7 +143,7 @@ class _AppState extends State<App> {
               fullscreen: false,
               overlayColor: Colors.blue.withOpacity(0.9),
               overlayShape: const CircleBorder(),
-              overlayBehavior: HitTestBehavior.deferToChild,
+              overlayBehavior: OverlayBehavior.deferToChild,
               showPulseAnimation: true,
             ),
             OnboardingStep(
@@ -187,14 +184,17 @@ class _AppState extends State<App> {
               bodyText: 'You can open menu from here',
               overlayColor: Colors.green.withOpacity(0.9),
               shape: const CircleBorder(),
-              overlayBehavior: HitTestBehavior.translucent,
+              overlayBehavior: OverlayBehavior.deferToOverlay,
+              onTapCallback:
+                  (TapArea area, VoidCallback next, VoidCallback close) {
+                scaffoldKey.currentState?.openDrawer();
+                next();
+              },
             ),
             OnboardingStep(
               focusNode: focusNodes[5],
-              titleText: 'Settings',
-              shape: const CircleBorder(),
-              bodyText:
-                  'Click here to access settings such as dark mode, daily limit, etc',
+              titleText: 'Close menu',
+              bodyText: 'Click here to close the drawer',
               fullscreen: false,
               overlayColor: Colors.black.withOpacity(0.8),
               overlayShape: const CircleBorder(),
@@ -209,6 +209,13 @@ class _AppState extends State<App> {
                 ),
               ),
               hasLabelBox: true,
+              hasArrow: true,
+              margin: EdgeInsets.all(0),
+              onTapCallback:
+                  (TapArea area, VoidCallback next, VoidCallback close) {
+                scaffoldKey.currentState?.openEndDrawer();
+                next();
+              },
             ),
             OnboardingStep(
               focusNode: focusNodes[6],
@@ -369,9 +376,9 @@ class _AppState extends State<App> {
           onChanged: (int index) {
             if (index == 4) {
               // close the drawer
-              if (scaffoldKey.currentState?.isDrawerOpen ?? false) {
-                scaffoldKey.currentState?.openEndDrawer();
-              }
+              // if (scaffoldKey.currentState?.isDrawerOpen ?? false) {
+              //   scaffoldKey.currentState?.openEndDrawer();
+              // }
               // interrupt onboarding on specific step
               // widget.onboardingKey.currentState.hide();
             }
@@ -437,7 +444,7 @@ class _HomeState extends State<Home> {
         ),
         actions: [
           IconButton(
-            focusNode: widget.focusNodes[5],
+            // focusNode: widget.focusNodes[5],
             icon: const Icon(Icons.mic),
             onPressed: () {},
           ),
@@ -453,9 +460,20 @@ class _HomeState extends State<Home> {
           )
         ],
       ),
-      drawer: const Drawer(
-        child: Center(
-          child: Text('Menu'),
+      drawer: Drawer(
+        child: ListView(
+          children: [
+            const DrawerHeader(
+              child: Text('Menu'),
+            ),
+            ListTile(
+              focusNode: widget.focusNodes[5],
+              title: const Text('Close menu'),
+              onTap: () {
+                Navigator.of(context).pop();
+              },
+            )
+          ],
         ),
       ),
       body: ListView.builder(