+++ lib/onboarding_overlay.dart
@@ -3,5 +3,6 @@ library onboarding_overlay;
 export 'src/label_painter.dart' show LabelPainter, ArrowPosition;
 export 'src/onboarding_widget.dart' show Onboarding, OnboardingState;
 export 'src/overlay_painter.dart' show OverlayPainter;
-export 'src/step.dart' show OnboardingStep, OnboardingStepRenderInfo;
+export 'src/step.dart'
+    show OnboardingStep, OnboardingStepRenderInfo, OverlayBehavior, TapArea;
 export 'src/stepper.dart' show OnboardingStepper;