+++ CHANGELOG.md
@@ -1,6 +1,17 @@
+## [3.0.0-pre.6]
+
+- fix: showPulseAnimation
+- add: demo add default theme textStyles
+- add: onTapCallback
+- remove: manualControl
+- remove: closeKey
+- change: replace HitTestBehavior with OverlayBehavior
+
+
 ## [3.0.0-pre.5]
+
 - add: a closeKey to control if we have a close button and ignore taps for next step if there is such widget and we tap within it's boundaries
-- change: rename manualControl to manualNextControl
+- change: rename manualControl to manualControl
 - fix: showPulseAnimation
 - add: demo add default theme textStyles